+++ pkg/assembler/backends/neo4j/src.go
@@ -197,6 +197,26 @@ func (e *srcNamespaceToName) IdentifiablePropertyNames() []string {
 }
 
 func (c *neo4jClient) Sources(ctx context.Context, sourceSpec *model.SourceSpec) ([]*model.Source, error) {
+
+	// fields: [type namespaces namespaces.namespace namespaces.names namespaces.names.name namespaces.names.tag namespaces.names.commit]
+	fields := getPreloads(ctx)
+	nameRequired := false
+	namespaceRequired := false
+	for _, f := range fields {
+		if f == "namespaces" {
+			namespaceRequired = true
+		}
+		if f == "namespaces.names" {
+			nameRequired = true
+		}
+	}
+
+	if !namespaceRequired && !nameRequired {
+		return c.sourcesType(ctx, sourceSpec)
+	} else if namespaceRequired {
+		return c.sourcesNamespace(ctx, sourceSpec)
+	}
+
 	session := c.driver.NewSession(neo4j.SessionConfig{AccessMode: neo4j.AccessModeRead})
 	defer session.Close()
 
@@ -305,7 +325,7 @@ func (c *neo4jClient) Sources(ctx context.Context, sourceSpec *model.SourceSpec)
 	return result.([]*model.Source), nil
 }
 
-func (c *neo4jClient) SourcesType(ctx context.Context, sourceSpec *model.SourceSpec) ([]*model.Source, error) {
+func (c *neo4jClient) sourcesType(ctx context.Context, sourceSpec *model.SourceSpec) ([]*model.Source, error) {
 	session := c.driver.NewSession(neo4j.SessionConfig{AccessMode: neo4j.AccessModeRead})
 	defer session.Close()
 
@@ -355,7 +375,7 @@ func (c *neo4jClient) SourcesType(ctx context.Context, sourceSpec *model.SourceS
 	return result.([]*model.Source), nil
 }
 
-func (c *neo4jClient) SourcesNamespace(ctx context.Context, sourceSpec *model.SourceSpec) ([]*model.Source, error) {
+func (c *neo4jClient) sourcesNamespace(ctx context.Context, sourceSpec *model.SourceSpec) ([]*model.Source, error) {
 	session := c.driver.NewSession(neo4j.SessionConfig{AccessMode: neo4j.AccessModeRead})
 	defer session.Close()
