+++ pkg/assembler/graphql/resolvers/package.resolvers.go
@@ -12,32 +12,5 @@ import (
 
 // Packages is the resolver for the packages field.
 func (r *queryResolver) Packages(ctx context.Context, pkgSpec *model.PkgSpec) ([]*model.Package, error) {
-	// fields: [type namespaces namespaces.namespace namespaces.names namespaces.names.name namespaces.names.versions
-	// namespaces.names.versions.version namespaces.names.versions.qualifiers namespaces.names.versions.qualifiers.key
-	// namespaces.names.versions.qualifiers.value namespaces.names.versions.subpath]
-	fields := getPreloads(ctx)
-	nameRequired := false
-	namespaceRequired := false
-	versionRequired := false
-	for _, f := range fields {
-		if f == "namespaces" {
-			namespaceRequired = true
-		}
-		if f == "namespaces.names" {
-			nameRequired = true
-		}
-		if f == "namespaces.names.versions" {
-			versionRequired = true
-		}
-	}
-
-	if namespaceRequired && nameRequired && versionRequired {
-		return r.Backend.Packages(ctx, pkgSpec)
-	} else if nameRequired {
-		return r.Backend.PackagesName(ctx, pkgSpec)
-	} else if namespaceRequired {
-		return r.Backend.PackagesNamespace(ctx, pkgSpec)
-	} else {
-		return r.Backend.PackagesType(ctx, pkgSpec)
-	}
+	return r.Backend.Packages(ctx, pkgSpec)
 }
