+++ pkg/assembler/graphql/resolvers/resolver.go
@@ -20,36 +20,8 @@ package resolvers
 // It serves as dependency injection for your app, add any dependencies you require here.
 
 import (
-	"context"
-
-	"github.com/99designs/gqlgen/graphql"
 	"github.com/guacsec/guac/pkg/assembler/backends"
 )
-
-func getPreloads(ctx context.Context) []string {
-	return getNestedPreloads(
-		graphql.GetOperationContext(ctx),
-		graphql.CollectFieldsCtx(ctx, nil),
-		"",
-	)
-}
-
-func getNestedPreloads(ctx *graphql.OperationContext, fields []graphql.CollectedField, prefix string) []string {
-	preloads := []string{}
-	for _, column := range fields {
-		prefixColumn := getPreloadString(prefix, column.Name)
-		preloads = append(preloads, prefixColumn)
-		preloads = append(preloads, getNestedPreloads(ctx, graphql.CollectFields(ctx, column.Selections, nil), prefixColumn)...)
-	}
-	return preloads
-}
-
-func getPreloadString(prefix, name string) string {
-	if len(prefix) > 0 {
-		return prefix + "." + name
-	}
-	return name
-}
 
 type Resolver struct {
 	Backend backends.Backend
